+++ perception/probabilistic_occupancy_grid_map/lib/fusion_policy/fusion_policy.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "probabilistic_occupancy_grid_map/fusion/single_frame_fusion_policy.hpp"
+#include "autoware/probabilistic_occupancy_grid_map/fusion_policy/fusion_policy.hpp"
 
 namespace fusion_policy
 {
