+++ perception/probabilistic_occupancy_grid_map/src/fusion/synchronized_grid_map_fusion_node.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "probabilistic_occupancy_grid_map/fusion/synchronized_grid_map_fusion_node.hpp"
+#include "synchronized_grid_map_fusion_node.hpp"
 
 #include "autoware/probabilistic_occupancy_grid_map/cost_value/cost_value.hpp"
 #include "autoware/probabilistic_occupancy_grid_map/utils/utils.hpp"
